Prevent excess domain event failure logging (#6)
**Summary:**
My previous commit made gomainevents noisier...perhaps a little too noisy. This reduces some of the noise and opts for emitting the original error that caused the deadletter event, only after the event has been added to the deadletter queue. 

NOTE: An improvement on this is to introduce the concept of an unavailable exception. However, that's a larger undertaking

**Test Plan:**
 - [x] Made an event fail, and made sure that event didn't make it to airbrake until the deadletter was triggered